// src/generated/prisma/models/User.ts
export type UserModel = runtime.Types.Result.DefaultSelection<Prisma.$UserPayload>
// ...
export type $UserPayload<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
  name: "User"
  objects: {
    carts: Prisma.$CartPayload<ExtArgs>[]
  }
  scalars: runtime.Types.Extensions.GetPayloadResult<{
    id: string
    email: string
    passwordHash: string
    name: string
    role: $Enums.Role
    createdAt: Date
    updatedAt: Date
  }, ExtArgs["result"]["user"]>
  composites: {}
}